Revert "folder structure (for dashboard)" (#14)
