+++ .github/workflows/pylint.yml
@@ -2,7 +2,6 @@ name: Pylint
 
 on:
   - push
-  - pull_request
 
 jobs:
   standards-check-py:
@@ -21,6 +20,4 @@ jobs:
         pip install pylint
     - name: Analysing the code with pylint
       run: |
-        pip install -r requirements.txt
-        cd bot
-        pylint --good-names `ls -d *.py -m | sed s/.py//g | sed 's/ //g' | tr -d '\n'` -d R0801,C0301,R0903,R0912 `ls -R|grep .py$|xargs`
+        pip install -r requirements.txt; pylint --good-names `ls -d *.py -m | sed s/.py//g | sed 's/ //g' | tr -d '\n'` -d R0801,C0301,R0903,R0912 `ls -R|grep .py$|xargs`
