+++ start.sh
@@ -1,4 +1,3 @@
-cd bot
 while true
 do
   python3.8 core.py
